refactor(store): Move test to utility mod
